unify types according to schema design policy (#76)

+++ crates/freeze/src/datasets/nonce_diffs.rs
@@ -9,7 +9,7 @@ use std::collections::HashMap;
 pub struct NonceDiffs {
     n_rows: u64,
     block_number: Vec<Option<u32>>,
-    transaction_index: Vec<Option<u64>>,
+    transaction_index: Vec<Option<u32>>,
     transaction_hash: Vec<Option<Vec<u8>>>,
     address: Vec<Vec<u8>>,
     from_value: Vec<U256>,
@@ -101,7 +101,7 @@ pub(crate) fn process_nonce_diff(
     };
     columns.n_rows += 1;
     store!(schema, columns, block_number, *block_number);
-    store!(schema, columns, transaction_index, Some(transaction_index as u64));
+    store!(schema, columns, transaction_index, Some(transaction_index as u32));
     store!(schema, columns, transaction_hash, transaction_hash.clone());
     store!(schema, columns, address, addr.as_bytes().to_vec());
     store!(schema, columns, from_value, from);
